Refacor: add,upateComment 분리
